requirement.txt
asgiref==3.2.10
backports.zoneinfo==0.2.1
Django==3.1
django-decouple==2.1
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
Pillow==7.2.0
psycopg2-binary==2.8.4
PyJWT==1.7.1
pytz==2022.2
sqlparse==0.4.2
